[DRUP-561] Create some abstraction for the subscription list builder.
